Read classes from label map
I grew tired of doing copy and paste from the label map to the script.

There is additional command line parameter to provide the path to the label map. The label map will be read once in the script, parsed into a dictionary and passed to the create_tf_example method.

The names are the keys in the dictionary and the IDs are the values.

All tests passed and tfrecords have been created without issues.